# Backend/schemas/openapi.yaml
openapi: 3.0.0
info:
  title: Book Database Api
  description: A simple API which provides access to a database of books and their authors.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://turboexhibit-diegosalsa-3030.codio-box.uk/api/v1'
    description: Development server on Codio
components:
  securitySchemes:
    jwtAuth:     
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  responses:
    NotFoundError:
      description: The specified resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedError:
      description: JWT is missing or invalid. Please log in using the /login endpoint. 
      content:
        application/json:
          schema:
            type: string
      required:
        - message
    SchemaValidationError:
      description: Invalid resource schema provided in request body.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PermissionError:
      description: Access level not sufficient for accessing requested resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerSideError:
      description: Something went wrong on the server side. If this keeps repeating, contact the admin.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
  parameters:
    page:
      in: query
      name: page
      schema:
        type: integer
        default: 0
      description: The number of the page of items to show. Each page is the size of "limit".
    limit:
      in: query
      name: limit
      schema:
        type: integer
        default: 5
      description: The numbers of items to return.
    order:
      in: query
      name: order
      schema:
        type: string
        default: "id"
      description: The property by which to sort the results.
...
paths:
...
  /author:
    get:
      tags:
        - Author
      summary:
        Get all author records.
      description:
        Get data of all authors from the database. Does not include the full contents (request individual authors for this).
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/order'
      responses:
        '200':
          description: Array of all author records from the database.
          headers:
            X-Total-Count:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: ./author.json#/definitions/authorGet
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      security:
        - jwtAuth: []
      tags:
        - Author
      summary:
        Add author data to the database.
      description:
        Add author data to the database. Returns the record ID.
      responses:
        '201':
          description: Added succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
        '400':
          $ref: '#/components/responses/SchemaValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403': 
          $ref: '#/components/responses/PermissionError'
        '500':
          $ref: '#/components/responses/ServerSideError'
      requestBody:
        description: All author data. 
        required: true
        content:
          application/json:
            schema:
              $ref: ./author.json#/definitions/authorAdd
  /author/{id}:
    get:
      tags:
        - Author
      summary:
        Get a single author resource from the database.
      description:
        Get an author resource from the database.
      responses:
        '200':
          description: Full data of the author.
          content:
            application/json:
              schema:
                $ref: ./author.json#/definitions/authorGet
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      security:
        - jwtAuth: []
      tags:
        - Author
      description:
        Modify data of a single author.
      responses:
        '204':
          description: Modified author resource.
        '400':
          $ref: '#/components/responses/SchemaValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerSideError'
      requestBody:
        description: Modifiable author data and author ID.
        required: true
        content:
          application/json:
            schema:
              $ref: ./author.json#/definitions/authorUpd
    delete:
      security:
        - jwtAuth: []
      tags:
        - Author
      description:
        Delete the data of a single author.
      responses:
        '204':
          description: Deleted author.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerSideError'
  /author/unapproved:
    get:
      tags:
        - Author
      summary:
        Get all unapproved author records. Admin only.
      description: >
        Get data of all authors from the database which are awaiting approval.
        Does not include the full contents (request individual authors for this).
        Admin only.
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/order'
      responses:
        '200':
          description: Array of all unapproved author records from the database.
          headers:
            X-Total-Count:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: ./author.json#/definitions/authorGet
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      security:
        - jwtAuth: []
      tags:
        - Author
      summary:
        Change the approval status of a author in the database.
      description: >
        Change the approval status of a author submitted to the database. 
        0 is default, unapproved and invisible to nonadmins.
        Send "approve" to approve the author and make it visible,
        "reject" to delete it from the database.
        Admin only.
      responses:
        '204':
          description: Modified succesfully.
        '400':
          $ref: '#/components/responses/SchemaValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403': 
          $ref: '#/components/responses/PermissionError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      requestBody:
        description: author ID and "approve" or "reject".
        required: true
        content:
          application/json:
            schema:
              $ref: ./author.json#/definitions/authorApprove
...